+++ src/api/components/files/files.controller.js
@@ -3,11 +3,14 @@
 
 const path = require('path');
 
-const { ConfigService } = require('../../../services/config/config.service');
-
 const SettingsModel = require('../settings/settings.model');
 
-exports.serve = async (req, res) => {
+const { ConfigService } = require('../../../services/config/config.service');
+const { LoggerService } = require('../../../services/logger/logger.service');
+
+const { log } = LoggerService;
+
+exports.serve = async (req, res, next) => {
   try {
     const recordingSettings = await SettingsModel.getByTarget(false, 'recordings');
 
@@ -23,7 +26,15 @@ exports.serve = async (req, res) => {
       root: path.join(recPath),
     };
 
-    res.sendFile(file, options);
+    res.sendFile(file, options, (err) => {
+      if (err) {
+        if (err?.status === 404 || err?.statusCode === 404) {
+          log.debug(err.message);
+        }
+
+        next();
+      }
+    });
   } catch (error) {
     res.status(500).send({
       statusCode: 500,